Use an in-memory DB instead of a real one.
This commit replaces the real MySQL DB uses in testing with a SQLite3 DB that is loaded from the spec/schema.rb file.  It's much more kind to the people who will use and maintain this in the future if the specs do not require DB setup for a developer to run them.

This commit uses Rails 3.2 because that is one of the two versions of Rails that this gem will need to be compatible with.  One of the two apps that will use this gem uses Rails 3.2 and Ruby 2.0.  The other is a legacy app that uses Ruby 1.8.7 and Rails 2.3.